B2-1 第九批: Net/Diagnostics/IO/Text/CSPRNG 链 calloc -> byte[]
Socket/AsyncSocket/TcpClient、Mqtt(Broker/Client)、WebSocket(WsFrame/WsReader/
WsWriter/Wss*)、Worker、TlsStream、WebApp 的 raw calloc 缓冲改成 ARC 管理的
byte[]，接收路径统一用 buf.ToStr(0, n)（长度来自 recv 返回值，不依赖 NUL）。
WsFrame.Encode() 原先 calloc 出去从没人释放，改成返回 byte[] 顺带修掉泄漏。

同批：ProcessHost/ServerMetrics 的 OS 探针缓冲、Directory 的 WIN32_FIND_DATAW、
Encoding 的临时缓冲与 Sha1(digest)、ResourcePack writer 侧收尾（nameHash/
MixKey/PackIndexEntry/storeBE* 全部 byte[]），以及
RandomNumberGenerator.GetBytes -> byte[] 及其 17 个调用点。
FbSrp.Create 的 FbBytes.Free 与 WXBizMsgCrypt.RandAscii 的 free 必须一起删，
否则 C free 掉 ARC 缓冲会堆损坏（0xC0000374）。

另修 gui_runtime_mac.m: zan_gui_wake 前向声明 i64 / 定义 i32 冲突，
Xcode clang 报错，drivers.yml 的 macOS GUI job 就卡在这里。

ctest 711/711 Passed。

+++ src/runtime/gui_runtime_mac.m
@@ -36,7 +36,7 @@ extern void zan_gui_internal_surface_clip(i64 id, int *x0, int *y0,
 
 /* Defined below; used by the client-side title-bar metrics helpers. */
 i32 zan_gui_get_dpi_scale(void);
-i64 zan_gui_wake(void);
+i32 zan_gui_wake(void);
 
 #define EXPORT __attribute__((visibility("default")))
